- **trying to approach universal level**
- **checking gltf**
- **some mobile stuff**
- **getting icloud storage path**
- **trying to store data in icloud**
- **better menu with conditional disabling**
- **on disk root path**
- **updated netrun**
- **started infinite scroll test, after setup refactoring**
- **reworked ui animations**
- **moved table layout and test to separate files**
- **double column table view first implementation**
- **new refs with dropping only on main**
- **layout table methods by ref**
- **disable drawing touches in ui tests**
- **fixed scroll events triggering**
- **basic scroll with double rows test**
- **complete infinite scroll test**
- **scroll view touch system**
- **infinite scroll scene on main screen**
- **ui animation with finish condition**
- **removed view touch internal trair**
- **debugging inertial scroll**
- **first implementation of inertial scrolling**
- **cleanup**

+++ deps/ui/src/input/touch.rs
@@ -7,7 +7,7 @@ use crate::{TouchLock, input::TouchEvent};
 
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub struct Touch {
-    pub id:       u64,
+    pub id:       usize,
     pub position: Point,
     pub event:    TouchEvent,
     pub button:   MouseButton,
